backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "chromadb>=1.4.1",
    "fastapi>=0.128.0",
    "groq>=1.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "ruff>=0.14.13",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
